Code refactoring, removed duplicate types
