+++ Microsoft.Toolkit.Mvvm/Messaging/Messenger.cs
@@ -8,6 +8,7 @@ using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using Microsoft.Collections.Extensions;
+using Microsoft.Toolkit.Mvvm.Messaging.Internals;
 
 namespace Microsoft.Toolkit.Mvvm.Messaging
 {
@@ -30,8 +31,8 @@ namespace Microsoft.Toolkit.Mvvm.Messaging
         //                    | /                       /                                            \
         // DictionarySlim<Recipient, DictionarySlim<TToken, MessageHandler<TRecipient, TMessage>>> mapping = Mapping<TMessage, TToken>
         //                                            /                                   /          /
-        //                      ___(Type2.tToken)____/                                   /          /
-        //                     /________________(Type2.tMessage)________________________/          /
+        //                      ___(Type2.TToken)____/                                   /          /
+        //                     /________________(Type2.TMessage)________________________/          /
         //                    /       ____________________________________________________________/
         //                   /       /
         // DictionarySlim<Type2, IMapping> typesMap;
@@ -590,80 +591,6 @@ namespace Microsoft.Toolkit.Mvvm.Messaging
             public override int GetHashCode()
             {
                 return RuntimeHelpers.GetHashCode(this.Target);
-            }
-        }
-
-        /// <summary>
-        /// A simple type representing an immutable pair of types.
-        /// </summary>
-        /// <remarks>
-        /// This type replaces a simple <see cref="ValueTuple{T1,T2}"/> as it's faster in its
-        /// <see cref="GetHashCode"/> and <see cref="IEquatable{T}.Equals(T)"/> methods, and because
-        /// unlike a value tuple it exposes its fields as immutable. Additionally, the
-        /// <see cref="tMessage"/> and <see cref="tToken"/> fields provide additional clarity reading
-        /// the code compared to <see cref="ValueTuple{T1,T2}.Item1"/> and <see cref="ValueTuple{T1,T2}.Item2"/>.
-        /// </remarks>
-        private readonly struct Type2 : IEquatable<Type2>
-        {
-            /// <summary>
-            /// The type of registered message.
-            /// </summary>
-            private readonly Type tMessage;
-
-            /// <summary>
-            /// The type of registration token.
-            /// </summary>
-            private readonly Type tToken;
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="Type2"/> struct.
-            /// </summary>
-            /// <param name="tMessage">The type of registered message.</param>
-            /// <param name="tToken">The type of registration token.</param>
-            [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public Type2(Type tMessage, Type tToken)
-            {
-                this.tMessage = tMessage;
-                this.tToken = tToken;
-            }
-
-            /// <inheritdoc/>
-            [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public bool Equals(Type2 other)
-            {
-                // We can't just use reference equality, as that's technically not guaranteed
-                // to work and might fail in very rare cases (eg. with type forwarding between
-                // different assemblies). Instead, we can use the == operator to compare for
-                // equality, which still avoids the callvirt overhead of calling Type.Equals,
-                // and is also implemented as a JIT intrinsic on runtimes such as .NET Core.
-                return
-                    this.tMessage == other.tMessage &&
-                    this.tToken == other.tToken;
-            }
-
-            /// <inheritdoc/>
-            public override bool Equals(object? obj)
-            {
-                return obj is Type2 other && Equals(other);
-            }
-
-            /// <inheritdoc/>
-            [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public override int GetHashCode()
-            {
-                unchecked
-                {
-                    // To combine the two hashes, we can simply use the fast djb2 hash algorithm.
-                    // This is not a problem in this case since we already know that the base
-                    // RuntimeHelpers.GetHashCode method is providing hashes with a good enough distribution.
-                    int hash = RuntimeHelpers.GetHashCode(this.tMessage);
-
-                    hash = (hash << 5) + hash;
-
-                    hash += RuntimeHelpers.GetHashCode(this.tToken);
-
-                    return hash;
-                }
             }
         }
